crypto/cipher: xorBytes performance ppc64le/ppc64
Existing code in xor.go checks for GOARCH values of 386 and
amd64 when doing efficient aligned xors of two buffers, and other
platform do xors of a byte at a time.  On Power (ppc64 & ppc64le)
this is very inefficient, and if the buffers being used to
do the xor are aligned the same, then the more efficient
xor loop can be done.  This adds a new function to check the
alignment of all buffers and then uses the efficient xor loop.

This change checks for GOARCH values of ppc64le and ppc64, but
other platforms might be able to take advantage of this too.

Fixes #14350

Change-Id: Iccc6b9d3df2e604a55f4c1e4890bdd3bb0d77ab0