[dev.garbage] runtime: replace ref with allocCount
This is a renaming of the field ref to the
more appropriate allocCount. The field
holds the number of objects in the span
that are currently allocated. Some throws
strings were adjust to more accurately
convey the meaning of allocCount.

I also removed some duplicate and uneeded
"const bitsPerByte 8" statements.

No logic or code paths were change with this
CL.

Change-Id: I10daf44e3e9cc24a10912638c7de3c1984ef8efe