net: Allow netgo to use lookup from nsswtich.conf
Commit 4a0ba7a allowed golang to use its own dns resolver instead of
libc in a number of cases. The code parses nsswitch.conf and attempts to
resolve things in the same order. Unfortunately, builds with netgo
completely ignore this parsing and always search via hostLookupFilesDNS.

This commit modifies the logic to allow binaries built with netgo to
parse nsswitch.conf and attempt to resolve using the order specified
there. If the parsing results in hostLookupCGo, it falls back to the
original hostLookupFilesDNS. Tests are also added to ensure that both
the parsing and the fallback work properly.

Fixes #14354

Change-Id: Ib079ad03d7036a4ec57f18352a15ba55d933f261