Update patch set 1
Patch Set 1: Run-TryBot+1

(1 comment)

I am in favor of this change but I am worried about unintended consequences (which I don't see at the moment).

Note that import is a file-specific operation - a package imported in one file is not visible in another. Now, the runtime package is never visible. Similarly, when importing unsafe, only when actually doing the operation importfile does it become visible.

Thus it all seems pretty well. Are you confident that this won't introduce bugs due to some global state changes that we don't see?

Patch-set: 1