[dev.ssa] cmd/compile/internal/ssa: remove proven redundant controls.
* It does very simple bounds checking elimination:
for i := range a { foo(a[i]) }.
* Improves on the following redundant expression:
return a6 || (a6 || (a6 || a4)) || (a6 || (a4 || a6 || (false || a6)))

I patched in CL 12960 that does bounds, nil and constant propagation
to make sure this CL is not just redundant. Size of pkg/tool/linux_amd64/*
(excluding compile which is affected by this change):

-this -12960 95941968
+this -12960 95655048 -286920
-this +12960 95585264 -69784
+this +12960 95518744 -66520

While there is some overlap ~20% of the reduction is unique to the new pass.

Change-Id: I98957427399145fb33693173fd4d5a8d71c7cc20