Update patch set 8
Patch Set 8:

Wow, lots of earlier comments on this CL. Thanks everyone.

I agree with Nigel that the new assembly routines deserve simple tests of their own. End-to-end tests are great too, but the precision of function-specific tests is especially helpful for long-term maintenance. Assembly breaks easily (maybe my assembler changes will start emitting the wrong instruction, or maybe there will be a typo in a cleanup, and so on); having a test that says "hey this assembly is the problem!" is important.

Because the input space here is fairly limited, the tests should be fairly simple to write.

Thanks.

Patch-set: 8
Label: Code-Review=0