Update patch set 3
Patch Set 3:

At first i only wanted to write a path to remove the math package dependency. Then i noticed that the tests for positive/negative numbers are not consistent. In one code path the check is for num[0] == "+" in another its v >= 0. It seemed better to rely on the sign in num. Covering all the exit points of the function with a call to execute a simple defer function seems a bit heavy if we can just guard the one return where its needed in the new if arrangement.

Patch-set: 3
Label: Code-Review=0