Update patch set 8
Patch Set 8:

(9 comments)

>What did you try?

First version was a simple one byte/loop.

My first attempt was to decouple the pipeline and load 8 bytes, copy it, offset one by 4 bytes, copy each, ANDQ with 255 and INCL. Shift each by 8, repeat.  In theory this should use the pipeline better, but I found the speed to be equivalent, so I dropped it.

This was proposed on github by another user, and showed a consistent 2.5% speedup when used, so I decided to use this instead.

I will try to do an alternative version, so you guys can see if I am doing something silly.

I have fixed what I was comfortable fixing in the inline editor (comments). I may have to wait until Monday for the code changes to get in, unless I can get this to work on my home computer.

Patch-set: 8
Label: Code-Review=0