Update patch set 9
Patch Set 9:

(22 comments)

There's still a lot to this CL. Here's a first pass for now.

The new golden tests are great, but make this large CL even larger. How hard is it to split them off as a separate CL? Doing so would have the bonus that it's very clear then that the same tests pass both before and after your optimizations. Sure, I'm 99% certain that you know what you're doing in the process of upstreaming all of your changes, but we might as well eliminate any chance of human error, if doing so is not too hard.

Similarly, I still think that the asm functions (and its new test file) can be split off as a separate CL, especially if they're not actually used until Part 2. It should be really easy to move them to a separate "git codereview change", as each file either splits into one CL or the other. The benefit is that smaller yet still self-contained changes are easier to review and manage. However, if you and Russ have already privately agreed to a particular breakdown of work into CLs and consider such a re-org to be too fiddly, I'm happy to yield.

Patch-set: 9
Label: Code-Review=0