Update patch set 6
Patch Set 6:

If you want to do this (I am mildly supportive) it needs to be done much more thoroughly. You need to verify the signature as well as the name or you are likely to get false positives; consider a function Testify().

Also the name should be test.go, in line with the others. The printf test is not called malformed_printf.go. We're checking the code; it's assumed we'll complain about bad cases.

Patch-set: 6