Update patch set 7
Patch Set 7:

I think Andrew's  point was the most important. This shouldn't be a new test/file but instead should be incorporated into example.go, which checks the form of Examples. This is the same thing, really, and should coexist with that check.

I would be happy to do that myself.

Patch-set: 7