Update patch set 2
Patch Set 2: Code-Review-1

Interesting trick, but I don't want to go down this route.
First, this makes the compiler nondeterministic, as results depend on allocation layout which can change from run to run.
The scheme would totally break if we ever get a moving collector.
Also, reflect.Pointer() is the super-ugly underbelly of reflect (along with InterfaceData, UnsafeAddr, and the SliceHeader/StringHeader types).  It isn't pointer-safe, so it's very fragile.

Patch-set: 2
Label: Code-Review=-1