Update patch set 2
Patch Set 2:

Oh, now I get it. 

pprof says that the time spent in the make line in Decode is 75% of the total execution time. Not 99%, but high nonetheless. 

I can lower it to ~35% by using longer strings (like Decoding "hello world"x10 instead of "hello world"). 

Want me to do that?

Patch-set: 2