[dev.ssa] cmd/compile: reduce line number churn in generated code
In tighten and regalloc, make moved and inserted
instructions use the line number of their *use*, not their
*source*.  This reduces the tendency of debugger stepping
to "jump around" the program.

For register allocation this looks like a relatively safe
transformation.  For tighten, notice that instructions with
memory operands don't get moved; this ought to prevent
anything exciting from getting the "wrong" line number.

Change-Id: I59e2eeac4dca9168d8af3a93effbc5bdacac2881