Update patch set 3
Patch Set 3:

> > Some of the code is in the runtime cgo support CL 19805.
 > > crosscall2 will update SB.
 > > https://go-review.googlesource.com/#/c//1/src/runtime/cgo/asm_mips64x.s
 > 
 > Oh, I get it now,  this is requiring not only that the total
 > text/data size be less than 2G, but also that the program exist
 > entirely within a single 4G memory area.  I'm worried that the
 > latter requirement will break when using shared libraries (cmd/go
 > -buildmode=c-shared).

Yes, this may be a problem. We might be able to solve it by setting SB to the start (or the end) of .text. This needs some work on the linker. Can we do it later when we get there?

 > 
 > > Do you think we should introduce a new register name alias for
 > R23 so that
 > > people won't accidentally clobber R23?
 > > I guess we can't use SB, how about RSB?
 > 
 > SGTM

Done in CL 19864.

Patch-set: 3
Reviewer: Gerrit User 9670 <9670@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0