Update patch set 1
Patch Set 1:

I'll spend more time playing with examples to see how I like it.

The "-x" case is, if you write something like

a := -x
b := x+x
f()
c := a*b

whether you want the stepping to go f;c or f;a;b;c which is what I think we'll do if tighten uses source line numbers (because a and b calcs will float past the call to f).

It might be wrong to change tighten, or maybe we should only do it for references to parameters.

Patch-set: 1