Update patch set 1
Patch Set 1:

Is there any followup benefit to this?

I kind of liked that the LFOO tokens are allocated from the surrogate halves range, so they're guaranteed to be disjoint from any possible legitimate single-rune tokens.  (Not that Go has any >=utf8.RuneSelf single-rune tokens, or is likely to ever add any.)

Patch-set: 1