Update patch set 12
Patch Set 12:

The issue with loop bounds checking is that you have reason very carefully about overflow, and you probably cannot do it in a single pass.  I'm reluctant to say "never", but the facts that need to be gathered are:

init_i is zero
phi_i is zero or larger and less than len(a)
since len(a) is a non-negative number and phi_i is (at least) one less than it, inc_i = phi_i + 1 will not overflow.  Because phi_i is non-negative, inc_i is also non-negative.

phi_i = phi(init_i, inc_i) is non-negative. (and this is the stable point of the analysis, more or less).

certain idioms that people write for unrolled loops are (technically) not overflow-proof.  For example

  for i:=0; i + 3 < len(a); i+= 4 { ... }

if i > maxint - 3, then i+3 overflows and the comparison to len(a) is useless.

Patch-set: 12