Update patch set 12
Patch Set 12: Code-Review+2

I think this is a better bet than the bounds check code I worked on, at least for now.  I can see places to tweak it to make it faster, but let's get it in and I'll tinker with the bounds stuff as time permits to see if I can make it pull its weight, otherwise we'll see if that matters more on Arm.

Patch-set: 12
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2