Update patch set 12
Patch Set 12:

I think we might be able to get in-loop bounds checks out in #loop-nesting-depth + 1 passes, not sure.  Suppose you run the algorithm once but ignore backedges on phis.  Remember all the information presented on backedges.
Run it again, using the stored information at phis.  If no backedges change, we're done.

Important thing is to not get too excited about detailed info on the backedges; you only want the minimum for bounds checks, i.e., "i >= 0".  I think that suffices for bounds checks (lets you know that i is not negative) and the less you try to know the faster it will converge.

Patch-set: 12