Update patch set 1
Patch Set 1:

(2 comments)

I like this a lot.

I actually have a pending CL that attempts exactly the same but goes a bit further: Most of the operator tokens are actually not needed anymore: I removed them all and subsumed them under a new one LOPER.

My version even moves LCOMM and LMUL under LOPER which makes some of the parsing functions for pointer and channel types cumbersome (need to test for LOPER and opcode), so I was thinking of moving those our of LOPER (but still give them precedence as you do). I also was thinking of using the 0 precedence indicator to say "no operator" to simplify bexpr.

But this seems simpler, at least in the parser. There may still be opportunity to remove tokens, but maybe it doesn't matter.

I'll have a look at this tomorrow in detail.

Patch-set: 1