Update patch set 1
Patch Set 1:

(1 comment)

Could the fix for the Copyright line that causes most of the changes be done separately?  It would then be easier to check that CL, and actually skim through the remaining differences.

Isn't the preferred style to end the line on end of sentence?  I've seen Rob mention that a couple of times, and it avoids the issue as well as cutting future diff ripple.

Patch-set: 1
Reviewer: Gerrit User 10961 <10961@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0