Update patch set 1
Patch Set 1: Run-TryBot+1

> Patch Set 1:
> 
> I thought the tiny allocator made tiny allocations pretty efficient already. I'm surprised this helps.
> 
> Does this reduce the allocation _count_ only, or any other useful metric (time, memory)?

I saw it reduce time a little (there's a vague sentence at the bottom of the CL description). I didn't paste in numbers because while it looked real, there's a lot of noise on my laptop and I don't have an analysis that I trust that I can run easily.

There's a high-level concept behind why this can be faster than leaving it for the GC: we know the lifetimes of the *LSym and their data are tied together, but the GC doesn't.

Patch-set: 1