Update patch set 12
Patch Set 12: Code-Review-1

This whole change looks way too complicated for that it needs to do.

Why does this support positive or negative signs at all?

Why does it care about the int size? A port number is a uint16. It always fit in an int, and it doesn't need positive or negative signs.

Patch-set: 12
Label: Code-Review=-1