Update patch set 3
Patch Set 3:

> > I would be in favor of a test.
 > 
 > I added a test.
 > 
 > > I was going to say if it's not worth
 > > adding a test, it's probably not worth fixing.
 > 
 > I don't follow this logic whatsoever.

It's a sarcastic way of saying it needs a test.

Put another way: look at the build dashboard and see happy "ok" across all columns. Everything looks fine. So what is this change fixing?

Or: "If it ain't broke, don't fix it."

Your job as a fixer is to also break it, to prove that your fix works.

Every fix needs to include a test that would fail previously, but pass now.

Patch-set: 3