Update patch set 1
Patch Set 1:

(1 comment)

Just double checking: did you discover this because of a helpful error message from the linker about a relocation's alignment being off? If not, I think that's worth fixing too, because this is a fun new side of relocations I haven't seen before.

(Maybe fun is not the word.)

Patch-set: 1