Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > Just double checking: did you discover this because of a helpful
 > error message from the linker about a relocation's alignment being
 > off? If not, I think that's worth fixing too, because this is a fun
 > new side of relocations I haven't seen before.
 > 
 > (Maybe fun is not the word.)

Well, sort of. I actually realized it would be a problem the moment I saw the CL (I have had lots of 'fun' with byte alignment over the past couple of months)! However there is a nice helpful error message in our s390x port that does let you know when you try to use a relocation which is not 2 byte aligned. It does show the problem and looks something like this:

cmd/go
main.init.1: go.string.alldata+3974688 is not 2-byte aligned
main.init.1: go.string.alldata+3974688 is not 2-byte aligned
main.buildCompiler.Set: go.string.alldata+3974688 is not 2-byte aligned
main.init.2: go.string.alldata+3974688 is not 2-byte aligned
...

I'm happy to try and add a more generic check. I'll take a look and upload another CL.

Ideally I'd add a test which increases the value of Minalign for a given platform and then checks that a simple program still links and executes correctly. Not sure of the best approach for this though as the way things are structured at the moment it would require recompiling the compiler. I think minux is looking at cache-aligning global variables so that might change the way this is implemented too.

Patch-set: 1
Label: Code-Review=0