Update patch set 1
Patch Set 1:

(3 comments)

This check adds a map allocation and a few comparisons for every struct in the code it inspects. Perhaps it could lazily allocate that map when it first sees a json struct tag, for the common case where there are no json tags present.

Rob recently wrote up guidelines for the checks that should be included in vet. They are here: http://tip.golang.org/src/cmd/vet/README

With regard to this change, I am satisfied that it addresses a real correctness issue and that the check is precise. What concerns me is the middle paragraph on "frequency." That is, does this issue arise in real code frequently enough to makde the check worth it?

Patch-set: 1