Update patch set 15
Patch Set 15:

(3 comments)

I agree with Nigel about splitting this up and I think it's beneficial for more careful review. For example, I'm not that comfortable with assembly, so I don't have much of an opinion when it comes to that, but I do feel very comfortable with the pure Go implementation of DEFLATE, so I can review that. Having a giant CL that mixes both components makes it harder for to be confident about a CL.

Patch-set: 15
Label: Code-Review=0