all: replace zero padding with precision when used with %#x
Change occurrences of integer formatting of the form "#0.*x" to
use a specific precision for formatting instead. These rely on
special behavior of width formatting when # and 0 padding is used
together that does not count the "0x" for the width. Which is
different from the usual padding behavior for numbers.

This change makes these occurrences keep their intended formatting
when a zero padding behavior change for # changes the
output of the original formatting specification.

Updates #14616

Change-Id: Ib4fb66f3882ed76321016d2cb4a89de8758e34c7